Add renderResume function to controllers folder.
