+++ controller/render/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  resume: require('./resume')
+};
